backend/requirements.txt
# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Cache
redis==5.0.1
aioredis==2.0.1

# AI and Vector Store
openai>=1.0.0
langchain==0.0.330
pinecone-client==2.2.4
sentence-transformers==2.2.2
huggingface-hub==0.16.4
torch>=1.6.0
transformers>=4.6.0
pymilvus>=2.3.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# Utils
websockets>=10.0
PyYAML==6.0.1
python-dotenv>=0.19.0
tenacity==8.2.3
structlog==23.2.0

# AWS
boto3==1.29.3
botocore==1.32.3
